util/sign-jwt-token/requirements.txt
cffi==1.17.1
cryptography==43.0.3
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
